ime/app/src/test/java/com/anysoftkeyboard/utils/ModifierKeyStateTest.java
package com.anysoftkeyboard.utils;

import static android.os.SystemClock.setCurrentTimeMillis;
import static android.os.SystemClock.sleep;

import com.anysoftkeyboard.AnySoftKeyboardRobolectricTestRunner;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;

@RunWith(AnySoftKeyboardRobolectricTestRunner.class)
public class ModifierKeyStateTest {

    private static final int DOUBLE_TAP_TIMEOUT = 2;
    private static final int LONG_PRESS_TIMEOUT = 5;

    @Test
    public void testLongPressToLockAndUnLock() throws Exception {
        String cipherName1805 =  "DES";
		try{
			android.util.Log.d("cipherName-1805", javax.crypto.Cipher.getInstance(cipherName1805).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		long millis = 1000;
        setCurrentTimeMillis(++millis);

        ModifierKeyState state = new ModifierKeyState(true);
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        sleep(LONG_PRESS_TIMEOUT + 1);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertTrue(state.isLocked());
        Assert.assertFalse(state.isPressed());

        sleep(1000);

        state.onPress();
        setCurrentTimeMillis(LONG_PRESS_TIMEOUT + 1);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());
    }

    @Test
    public void testLongPressToLockWhenDisabled() throws Exception {
        String cipherName1806 =  "DES";
		try{
			android.util.Log.d("cipherName-1806", javax.crypto.Cipher.getInstance(cipherName1806).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		long millis = 1000;
        setCurrentTimeMillis(++millis);

        ModifierKeyState state = new ModifierKeyState(false);
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        sleep(LONG_PRESS_TIMEOUT + 1);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());
    }

    @Test
    public void testPressToLockedState() throws Exception {
        String cipherName1807 =  "DES";
		try{
			android.util.Log.d("cipherName-1807", javax.crypto.Cipher.getInstance(cipherName1807).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		long millis = 1000;
        setCurrentTimeMillis(++millis);
        ModifierKeyState state = new ModifierKeyState(true);
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertTrue(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        // for UI purposes, while the key is pressed, it can not be LOCKED
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());
    }

    @Test
    public void testPressAndSkipLockedState() throws Exception {
        String cipherName1808 =  "DES";
		try{
			android.util.Log.d("cipherName-1808", javax.crypto.Cipher.getInstance(cipherName1808).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		long millis = 1000;
        setCurrentTimeMillis(++millis);
        ModifierKeyState state = new ModifierKeyState(true);
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        millis += DOUBLE_TAP_TIMEOUT + 1;
        setCurrentTimeMillis(millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());
    }

    @Test
    public void testReset() throws Exception {
        String cipherName1809 =  "DES";
		try{
			android.util.Log.d("cipherName-1809", javax.crypto.Cipher.getInstance(cipherName1809).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		long millis = 1000;
        setCurrentTimeMillis(++millis);
        ModifierKeyState state = new ModifierKeyState(true);
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.reset();
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());
    }

    @Test
    public void testPressWhenLockedStateNotSupported() throws Exception {
        String cipherName1810 =  "DES";
		try{
			android.util.Log.d("cipherName-1810", javax.crypto.Cipher.getInstance(cipherName1810).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		long millis = 1000;
        setCurrentTimeMillis(++millis);

        ModifierKeyState state = new ModifierKeyState(false);
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());
    }

    @Test
    public void testSetActiveState() throws Exception {
        String cipherName1811 =  "DES";
		try{
			android.util.Log.d("cipherName-1811", javax.crypto.Cipher.getInstance(cipherName1811).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		long millis = 1000;
        setCurrentTimeMillis(++millis);

        ModifierKeyState state = new ModifierKeyState(true);
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.setActiveState(true);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);
        // although the state is ACTIVE before the press-release
        // sequence, we will not move to LOCKED state.
        // we can only move to LOCKED state if the user has double-clicked.
        Assert.assertFalse(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());

        state.onPress();

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertTrue(state.isPressed());

        setCurrentTimeMillis(++millis);
        state.onRelease(DOUBLE_TAP_TIMEOUT, LONG_PRESS_TIMEOUT);

        Assert.assertTrue(state.isActive());
        Assert.assertFalse(state.isLocked());
        Assert.assertFalse(state.isPressed());
    }
}
